students/k3339/Malakhov_Alexey/Lr3/fastapi/pyproject.toml
[project]
name = "malakhov-alexey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "fastapi[all]>=0.135.3",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.37",
]

[tool.basedpyright]
reportReturnType = false
